+++ src/command.ts
@@ -1,7 +1,8 @@
 import { window, workspace, TextEditor, Range } from "vscode";
-import { JSXElement, JSXIdentifier } from "@babel/types";
-import { parseDocument, IStyleAttribute } from "./parsing";
-import { generateStyledComponent } from "./generation";
+import { JSXElement } from "@babel/types";
+import { parseDocument, IStyleAttribute } from "./util/parseDocument";
+import { generateStyledComponent } from "./util/generateStyledComponent";
+import { generateImportStatement } from "./util/generateImportStatement";
 
 export const COMMAND_NAME = "extension.styco";
 
@@ -15,35 +16,41 @@ export const stycoCommand = async () => {
     return;
   }
 
-  const documentInformation = parseDocument(editor);
-  if (!documentInformation) {
-    window.showInformationMessage("Could not find element or style attribute");
-    return;
-  }
+  const documentInformation = parseDocument(
+    editor.document.getText(),
+    editor.document.offsetAt(editor.selection.active)
+  );
 
   const {
     selectedElement,
     elementName,
     insertPosition,
-    styleAttr
+    styleAttr,
+    importStatementExisting,
   } = documentInformation;
 
   const stycoName = await window.showInputBox({
     prompt: "Name: ",
-    placeHolder: "Name of the component"
+    placeHolder: "Name of the component",
   });
 
   if (!stycoName) {
     window.showInformationMessage("Please enter a name");
     return;
   }
-
   const component = generateStyledComponent(elementName, stycoName, styleAttr);
+
+  const importStatement =
+    importStatementExisting ||
+    workspace.getConfiguration("styco").get("insertImportStatement") === false
+      ? null
+      : await generateImportStatement(editor.document.uri);
 
   try {
     await modifyDocument(
       editor,
       component,
+      importStatement,
       insertPosition,
       selectedElement,
       styleAttr,
@@ -53,7 +60,6 @@ export const stycoCommand = async () => {
     window.showInformationMessage("Could not update document");
     return;
   }
-
   if (workspace.getConfiguration("styco").get("saveAfterExecute")) {
     await editor.document.save();
   }
@@ -62,25 +68,37 @@ export const stycoCommand = async () => {
 const modifyDocument = async (
   editor: TextEditor,
   styledComponent: string,
+  importStatement: string | null,
   insertPosition: number,
   oldElement: JSXElement,
   styleAttr: IStyleAttribute | null,
   stycoName: string
 ) => {
   const { document } = editor;
-  const openName = oldElement.openingElement.name as JSXIdentifier;
-  const closeName =
-    oldElement.closingElement !== null
-      ? (oldElement.closingElement.name as JSXIdentifier)
-      : null;
+  const openName = oldElement.openingElement.name;
+  const closeName = oldElement.closingElement?.name;
 
   await editor.edit(
     editBuilder => {
-      // Insert StyCo
-      editBuilder.insert(
-        editor.document.positionAt(insertPosition),
-        `\n\n${styledComponent}`
-      );
+      // Insert import statement
+      if (importStatement !== null) {
+        editBuilder.insert(
+          document.positionAt(insertPosition),
+          `\n${importStatement}`
+        );
+
+        // Insert StyCo below
+        editBuilder.insert(
+          document.positionAt(insertPosition + 1),
+          `\n${styledComponent}\n`
+        );
+      } else {
+        // Insert StyCo
+        editBuilder.insert(
+          document.positionAt(insertPosition),
+          `\n\n${styledComponent}\n`
+        );
+      }
 
       // Remove style-attribute
       if (styleAttr !== null) {
@@ -102,7 +120,7 @@ const modifyDocument = async (
       );
 
       // Rename Closing Tag
-      if (closeName !== null) {
+      if (closeName !== undefined) {
         editBuilder.replace(
           new Range(
             document.positionAt(closeName.start!),